Updated offline file structure.
