+++ src/components/choose-locale/change-locale-offline.component.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { postUserPropertiesOffline } from '../../offline/user-properties.resource';
+import { postUserPropertiesOffline } from './change-locale-offline.resource';
 import ChangeLocale, { ChangeLocaleProps } from './change-locale.component';
 
 const ChangeLocaleOffline: React.FC<ChangeLocaleProps> = props => (
