+++ src/components/choose-locale/change-locale-online.resource.tsx
@@ -1,0 +1,14 @@
+import { openmrsFetch } from '@openmrs/esm-framework';
+
+export function postUserPropertiesOnline(
+  userUuid: string,
+  userProperties: any,
+  abortController?: AbortController,
+): Promise<any> {
+  return openmrsFetch(`/ws/rest/v1/user/${userUuid}`, {
+    method: 'POST',
+    body: { userProperties: userProperties },
+    headers: { 'Content-Type': 'application/json' },
+    signal: abortController?.signal,
+  });
+}
