+++ src/components/choose-locale/change-locale.resource.tsx
@@ -1,0 +1,5 @@
+export type PostUserProperties = (
+  userUuid: string,
+  userProperties: any,
+  abortController?: AbortController,
+) => Promise<any>;
