+++ src/root.component.tsx
@@ -5,7 +5,7 @@ import { BrowserRouter, Redirect } from 'react-router-dom';
 import { getCurrentUser, createErrorHandler } from '@openmrs/esm-framework';
 import { LoggedInUser, UserSession } from './types';
 import { getCurrentSession } from './root.resource';
-import { syncUserPropertiesChanges } from './offline/user-properties.resource';
+import { syncUserPropertiesChanges } from './offline';
 
 export interface RootProps {
   syncUserPropertiesChangesOnLoad: boolean;
